Implement single-line arc-list packing optimization
If the cell only has one active arc, and that arc is
in fact a line, then encode the line in the cell entry
itself.  This reduces total texture fetches for the cell
from 3 to 1.  It's fair to say that this reduces total
averate texture fetches of average fonts by at least .5.