Add more stable distance code for extremely shallow arcs
For shallow arcs, using the arc center to do the distance calculations
was too unstable, resulting in seams that were going wild.  This was
most obvious with Shojumaru-Regular.ttf small letter 'h'.

This is what the new code does.  For arcs with abs(d) < .01, we
make some sin(a)=tan(a)=a assumptions and approximate the distance.

This essentially turns the arc into a conic.  This is problematic
in the inside side of the wedge as now the arc corners are closer than
the arc.  So we get the corner distances too, and return the minimum
of all.

This should resolve most seam glitches.