Fix tricky abs bug
Wojciech and I observed that gyphy misrenders if glyph geometry is very
small.  I reproduced by dividing all geometry by upem in glyphy-demo.

The bug was simple, an innocient use of abs() instead of fabs() was
rounding distances.  This probably happened when I copied the sdf logic
from GPU to the CPU.

Also improve the epsilon handling there to be relative, not absolute.
Both on CPU and GPU.

There still are srtifacts showing with scaled-down glyph.  Debugging.