Add perspective transformations
Perspective can be changed using ctrl+right-button.

The perspective-correct interpolation of varying variables loses a bit
or two, so I had to reduce nominal_size from 8 bits to 6 bits and spare
the two lowest-significant bits of the varying to get correct results.
In OpenGL 3 and later we can use "flat in/out" instead of varying to
disable interpolation.