meson.build: Fix usage as subprojects
It was pointed out that we might fall into issues with the FreeType dependency,
so do what we did before.  That is, look for the 'freetype2' dependency using
pkg-config, before we look for the 'Freetype' dependency using CMake on Visual
Studio-like compilers.

We can do this in one single dependency line in the future, when we require
Meson 0.60.0 or later.