Add Github Actions CI Workflow to build/test on Ubuntu 22.04
Builds the library, demo, and runs the validator.

meson.build has been updated to capitalize the GLUT dependency name
so it is detected correctly by Meson.

Note: Github CI runners are Ubuntu 22.04, which has an old version
of harfbuzz, so harfbuzz must be built from source. The harfbuzz
build can be removed when Github updates the CI runners to
Ubuntu 24.04.